+++ Modules/Utils/Sources/PocketCastsUtils/Feature Flags/FeatureFlag.swift
@@ -14,12 +14,6 @@ public enum FeatureFlag: String, CaseIterable {
     /// Whether End Of Year feature is enabled
     case endOfYear
 
-    /// Enable show notes using the new endpoint
-    case newShowNotesEndpoint
-
-    /// Enable retrieving episode artwork from the RSS feed
-    case episodeFeedArtwork
-
     /// Enable chapters to be loaded from the RSS feed
     case rssChapters
 
@@ -164,10 +158,6 @@ public enum FeatureFlag: String, CaseIterable {
             false
         case .endOfYear:
             false
-        case .newShowNotesEndpoint:
-            false
-        case .episodeFeedArtwork:
-            false
         case .rssChapters:
             false
         case .errorLogoutHandling:
@@ -263,10 +253,6 @@ public enum FeatureFlag: String, CaseIterable {
             shouldEnableSyncedSettings ? "new_settings_storage" : nil
         case .settingsSync:
             shouldEnableSyncedSettings ? "settings_sync" : nil
-        case .newShowNotesEndpoint:
-             "new_show_notes"
-         case .episodeFeedArtwork:
-             "episode_artwork"
          case .rssChapters:
              "rss_chapters"
         case .categoriesRedesign: